Rework how update interval works
I'd must have misunderstood how to properly implement the update
interval mechanism, and limited it by a Throttle, and not by informing
how often to call update.

This caused the homeassistant.update_entity service not to be able to
force a refresh of these entities created, because the Throttle blocked
them.

This fixes so the update interval is informed to EntityComponent and
removes the Throttle, thus homeassistant.update_entity works.